feat(mediator): Remove a dependencia do MeadiaR e add uma novo

+++ src/EmpregaNet.Application/Common/Base/BaseHandlers.cs
@@ -1,9 +1,9 @@
 using EmpregaNet.Application.Common.Command;
+using EmpregaNet.Application.Messages;
 using EmpregaNet.Domain;
 using EmpregaNet.Domain.Common;
 using EmpregaNet.Domain.Interfaces;
 using FluentValidation;
-using MediatR;
 using Microsoft.Extensions.Logging;
 
 namespace EmpregaNet.Application.Common.Base;
@@ -14,12 +14,15 @@ public class CreateHandler<TEntity> : IRequestHandler<CreateCommand<TEntity, TEn
     protected readonly IBaseRepository<TEntity> _repository;
     protected readonly IValidator<CreateCommand<TEntity, TEntity>> _validator;
     protected readonly ILogger<CreateHandler<TEntity>> _logger;
+    protected readonly IMediator _mediator;
 
     public CreateHandler(
+        IMediator mediator,
         IBaseRepository<TEntity> repository,
         IValidator<CreateCommand<TEntity, TEntity>> validator,
         ILogger<CreateHandler<TEntity>> logger)
     {
+        _mediator = mediator;
         _repository = repository;
         _validator = validator;
         _logger = logger;
@@ -33,6 +36,9 @@ public class CreateHandler<TEntity> : IRequestHandler<CreateCommand<TEntity, TEn
         var createdEntity = await _repository.CreateAsync(request.entity);
 
         _logger.LogInformation("{EntityName} criada com sucesso. ID: {Id}", typeof(TEntity).Name, createdEntity.Id);
+
+
+        await _mediator.Publish(new EntityEvent<TEntity>(createdEntity), cancellationToken);
         return createdEntity;
     }
 
@@ -55,12 +61,15 @@ public class UpdateCommandHandler<TRequest, TResponse>
     private readonly IBaseRepository<TResponse> _repository;
     protected readonly IValidator<UpdateCommand<TRequest, TResponse>> _validator;
     private readonly ILogger<UpdateCommandHandler<TRequest, TResponse>> _logger;
+    protected readonly IMediator _mediator;
 
     public UpdateCommandHandler(
+        IMediator mediator,
         IValidator<UpdateCommand<TRequest, TResponse>> validator,
         IBaseRepository<TResponse> repository,
         ILogger<UpdateCommandHandler<TRequest, TResponse>> logger)
     {
+        _mediator = mediator;
         _logger = logger;
         _repository = repository;
         _validator = validator;
@@ -83,6 +92,8 @@ public class UpdateCommandHandler<TRequest, TResponse>
 
         var updatedEntity = await _repository.UpdateAsync(entity);
         _logger.LogInformation("{EntityName} atualizada com sucesso. ID: {Id}", typeof(TResponse).Name, updatedEntity.Id);
+        await _mediator.Publish(new EntityEvent<TResponse>(updatedEntity), cancellationToken);
+
         return updatedEntity;
     }
 
@@ -124,31 +135,59 @@ public sealed class GetByIdHandler<TEntity> : IRequestHandler<GetByIdQuery<TEnti
 
     public async Task<TEntity> Handle(GetByIdQuery<TEntity> request, CancellationToken cancellationToken)
     {
-        return await _repository.GetByIdAsync(request.Id) ??
-            throw new KeyNotFoundException($"{typeof(TEntity).Name} com ID {request.Id} não encontrado");
+        return await _repository.GetByIdAsync(request.Id) ?? throw new KeyNotFoundException($"{typeof(TEntity).Name} com ID {request.Id} não encontrado");
     }
 }
 
-public sealed class DeleteHandler<TEntity> : IRequestHandler<DeleteCommand<TEntity>>
+public sealed class DeleteHandler<TEntity, TResponse> : IRequestHandler<DeleteCommand<TResponse>, TResponse>
     where TEntity : BaseEntity
+    where TResponse : class
 {
     private readonly IBaseRepository<TEntity> _repository;
-    private readonly ILogger<DeleteHandler<TEntity>> _logger;
+    private readonly ILogger<DeleteHandler<TEntity, TResponse>> _logger;
+    private readonly IMediator _mediator;
 
     public DeleteHandler(
+        IMediator mediator,
         IBaseRepository<TEntity> repository,
-        ILogger<DeleteHandler<TEntity>> logger)
+        ILogger<DeleteHandler<TEntity, TResponse>> logger)
     {
+        _mediator = mediator;
         _repository = repository;
         _logger = logger;
     }
 
-    public async Task Handle(DeleteCommand<TEntity> request, CancellationToken cancellationToken)
+#pragma warning disable CS8613 
+    public async Task<TResponse?> Handle(DeleteCommand<TResponse> request, CancellationToken cancellationToken)
+#pragma warning restore CS8613
     {
         _logger.LogInformation("Removendo a entidade {EntityName} com o ID: {Id}",
             typeof(TEntity).Name, request.Id);
 
+        var entityToDelete = await _repository.GetByIdAsync(request.Id);
+        if (entityToDelete == null)
+        {
+            throw new KeyNotFoundException($"{typeof(TEntity).Name} com ID {request.Id} não encontrado");
+        }
+
         await _repository.DeleteAsync(request.Id);
         _logger.LogInformation("{EntityName} removido com sucesso. ID: {Id}", typeof(TEntity).Name, request.Id);
+        await _mediator.Publish(new EntityEvent<TEntity>(entityToDelete), cancellationToken);
+
+        if (typeof(TResponse) == typeof(object) || typeof(TResponse) == typeof(void)) // 'void' aqui é apenas um marcador, na prática será 'object'
+        {
+            return null;
+        }
+        else if (typeof(TResponse) == typeof(bool))
+        {
+            return (TResponse)(object)true;
+        }
+        else if (typeof(TResponse) == typeof(TEntity))
+        {
+            return (TResponse)(object)entityToDelete;
+        }
+
+
+        return default(TResponse)!;
     }
 }
